Rename jwt package, because it's not just used for jwt
