+++ tests/integration/pkg/auth/hydra.go
@@ -1,20 +1,18 @@
-package jwt
+package auth
 
 import (
 	"context"
 	"crypto/tls"
 	"fmt"
-	"net/http"
-	"net/http/cookiejar"
-	"net/url"
-	"time"
-
 	"golang.org/x/net/publicsuffix"
 	"golang.org/x/oauth2"
 	"golang.org/x/oauth2/clientcredentials"
+	"net/http"
+	"net/http/cookiejar"
+	"time"
 )
 
-func GetAccessToken(oauth2Cfg clientcredentials.Config, tokenType ...string) (string, error) {
+func GetHydraOAuth2AccessToken(oauth2Cfg clientcredentials.Config) (string, error) {
 	jar, err := cookiejar.New(&cookiejar.Options{PublicSuffixList: publicsuffix.List})
 	if err != nil {
 		return "", err
@@ -27,11 +25,6 @@ func GetAccessToken(oauth2Cfg clientcredentials.Config, tokenType ...string) (st
 		Jar:     jar,
 	}
 
-	if len(tokenType) > 0 {
-		oauth2Cfg.EndpointParams = make(url.Values)
-		oauth2Cfg.EndpointParams.Add("token_format", tokenType[0])
-	}
-
 	ctx := context.WithValue(context.Background(), oauth2.HTTPClient, httpClient)
 	token, err := oauth2Cfg.Token(ctx)
 	if err != nil {
@@ -40,8 +33,8 @@ func GetAccessToken(oauth2Cfg clientcredentials.Config, tokenType ...string) (st
 	if !token.Valid() {
 		return "", fmt.Errorf("token invalid. got: %#v", token)
 	}
-	if token.TokenType != "Bearer" {
-		return "", fmt.Errorf("token type = %q; want %q", token.TokenType, "Bearer")
+	if token.TokenType != "bearer" {
+		return "", fmt.Errorf("token type = %q; want %q", token.TokenType, "bearer")
 	}
 	return token.AccessToken, nil
 }
