+++ tests/integration/testsuites/istio-jwt/scenario.go
@@ -3,8 +3,8 @@ package istiojwt
 import (
 	"errors"
 	"fmt"
+	"github.com/kyma-project/api-gateway/tests/integration/pkg/auth"
 	"github.com/kyma-project/api-gateway/tests/integration/pkg/helpers"
-	"github.com/kyma-project/api-gateway/tests/integration/pkg/jwt"
 	"github.com/kyma-project/api-gateway/tests/integration/pkg/manifestprocessor"
 	"github.com/kyma-project/api-gateway/tests/integration/pkg/resource"
 	"github.com/kyma-project/api-gateway/tests/integration/pkg/testcontext"
@@ -90,7 +90,7 @@ func (s *scenario) callingEndpointWithHeadersWithRetries(url string, tokenType s
 		requestHeaders = make(map[string]string)
 	}
 
-	token, err := jwt.GetAccessToken(*s.oauth2Cfg, strings.ToLower(tokenType))
+	token, err := auth.GetAccessToken(*s.oauth2Cfg, strings.ToLower(tokenType))
 	if err != nil {
 		return fmt.Errorf("failed to fetch an id_token: %s", err.Error())
 	}
